cmake: enable -Wall -Werror -Wextra -march=native
I had thought these were already enabled but apparently they were not.
Lots of fixes to repair all the mistakes found by these warnings.

The correct way to add these options in cmake seems to be
the 'add_compile_options' directive.

That should also apply then to any subdirecties below where it is set.

[1] https://cmake.org/cmake/help/latest/command/add_compile_options.html#command

+++ compiler/main/x86_code_gen/compile_ir/compile_tac_label.c
@@ -24,7 +24,7 @@ bool compile_tac_label_x86(struct TAC* tac, struct IBuffer* ibu, struct Ctx* ctx
 
 		char* function_name;
 
-		if (tac_dest(tac) < sst_size(ctx_tables(ctx)->sst)) {
+		if (tac_dest(tac) < (int32_t)sst_size(ctx_tables(ctx)->sst)) {
 
 			function_name = sst_at(ctx_tables(ctx)->sst, tac_dest(tac))->name;
 		} else {
